+++ src/pages/orders-history.page.module.css
@@ -1,0 +1,7 @@
+.ordersList{
+  display: flex;
+  flex-flow: column;
+  gap: 16px;
+  overflow-y: scroll;
+  margin-bottom: var(--pageBottomMargin);
+}
